do not plot GWAS locuszoom plot if data file is not uploaded

+++ index.js
@@ -9,6 +9,8 @@ const rscript = require('./r-calculations/r-wrapper.js');
 
 var app = express();
 
+console.log("Starting server...");
+
 
 // Upload files with file extension and original name
 var storage = multer.diskStorage({
@@ -39,27 +41,45 @@ app.use(function(req, res, next) {
 
 app.post('/upload-file', upload.any(), async (request, response) => {
   console.log(request.files);
-  console.log('Files uploaded: ', request.files[0].filename, request.files[1].filename, request.files[2].filename, request.files[3].filename);
-  var associationFile = request.files[0].filename;
-  var expressionFile = request.files[1].filename;
-  var genotypeFile = request.files[2].filename;
-  var gwasFile = request.files[3].filename;
+  console.log("Files uploaded.");
+
+  var associationFile = request.files[0].filename; // required file
+  var expressionFile = 'false'; //optional data file
+  var genotypeFile = 'false'; //optional data file
+  var gwasFile = 'false'; //optional data file
+
+  if (request.files[1] !== undefined && request.files[2] !== undefined) {
+    expressionFile = request.files[1].filename;
+    genotypeFile = request.files[2].filename;
+  }
+  if (request.files[3] !== undefined) {
+    gwasFile = request.files[3].filename;
+  }
+
   var dir = __dirname + '/r-calculations/uploads/'
   try {
     const data = await rscript('./r-calculations/eQTL/eqtl.r', expressionFile, genotypeFile, associationFile, gwasFile);
     // remove files from uploads folder when data is received from R
     await removeFile(dir + associationFile);
-    await removeFile(dir + expressionFile);
-    await removeFile(dir + genotypeFile);
-    await removeFile(dir + gwasFile);
+    if (request.files[1] !== undefined && request.files[2] !== undefined) {
+      await removeFile(dir + expressionFile);
+      await removeFile(dir + genotypeFile);
+    }
+    if (request.files[3] !== undefined) {
+      await removeFile(dir + gwasFile);
+    }
     response.json(data);
   } catch(err) {
     console.log(err);
     // remove files from uploads folder when data is received from R
     await removeFile(dir + associationFile);
-    await removeFile(dir + expressionFile);
-    await removeFile(dir + genotypeFile);
-    await removeFile(dir + gwasFile);
+    if (request.files[1] !== undefined && request.files[2] !== undefined) {
+      await removeFile(dir + expressionFile);
+      await removeFile(dir + genotypeFile);
+    }
+    if (request.files[3] !== undefined) {
+      await removeFile(dir + gwasFile);
+    }
     response.status(500);
     response.json(err.toString());
   }
